webapp/requirements.txt
django>=4.2
openpyxl
requests
whitenoise
dj-database-url
psycopg2-binary
django-allauth
PyJWT
cryptography
httpx[http2]
django-cors-headers
